# packages/api/src/swagger/v0/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RegisterUserInput": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name"
				],
				"type": "object"
			},
			"UserRole": {
				"enum": [
					"SUPER_ADMIN",
					"ADMIN",
					"STUDENT",
					"TEACHER"
				],
				"type": "string"
			},
			"UserStatus": {
				"enum": [
					"PENDING",
					"TRIAL",
					"SUSPEND",
					"ACTIVE"
				],
				"type": "string"
			},
			"UserDto": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"role",
					"email",
					"name",
					"id"
				],
				"type": "object"
			},
			"LoginPayload": {
				"properties": {
					"isFirstLogin": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/UserDto"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"role": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT authorization token with role claim"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "0.1.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserInput"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginPayload"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"security": [
					{
						"role": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "GetAdminUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"role": [
							"SUPER_ADMIN"
						]
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/v0"
		}
	]
}